+++ examples/client/jwt-api/CHANGELOG.md
@@ -5,4 +5,4 @@
 ### Patch Changes
 
 - Updated dependencies [8b5f490]
-  - @openauthjs/openauth@0.2.4
+  - @kagii/openauth@0.2.4